Rename RoutingContext -> RouterContext
We only export `router` to context now and would
like to start a precedent that components that
export a context are named after what they export

closes #2650
