+++ modules/__tests__/serverRendering-test.js
@@ -2,7 +2,7 @@ import expect from 'expect'
 import React, { Component } from 'react'
 import { renderToString } from 'react-dom/server'
 import match from '../match'
-import RoutingContext from '../RoutingContext'
+import RouterContext from '../RoutingContext'
 import Link from '../Link'
 
 describe('server rendering', function () {
@@ -68,7 +68,7 @@ describe('server rendering', function () {
   it('works', function (done) {
     match({ routes, location: '/dashboard' }, function (error, redirectLocation, renderProps) {
       const string = renderToString(
-        <RoutingContext {...renderProps} />
+        <RouterContext {...renderProps} />
       )
       expect(string).toMatch(/The Dashboard/)
       done()
@@ -78,7 +78,7 @@ describe('server rendering', function () {
   it('renders active Links as active', function (done) {
     match({ routes, location: '/about' }, function (error, redirectLocation, renderProps) {
       const string = renderToString(
-        <RoutingContext {...renderProps} />
+        <RouterContext {...renderProps} />
       )
       expect(string).toMatch(/about-is-active/)
       expect(string).toNotMatch(/dashboard-is-active/)
